Tests/LibWeb/Text/expected/css/dir-pseudo-on-input-element.txt

Input matches :dir(ltr): true
Input matches :dir(rtl): false
Input matches :dir(ltr): true
Input matches :dir(rtl): false
Input matches :dir(ltr): false
Input matches :dir(rtl): true
Input matches :dir(ltr): true
Input matches :dir(rtl): false
Input matches :dir(ltr): false
Input matches :dir(rtl): true
Input matches :dir(ltr): true
Input matches :dir(rtl): false
Input matches :dir(ltr): true
Input matches :dir(rtl): false
